added null checks

+++ Core/BusinessLogic/src/main/java/org/onlinebanking/core/businesslogic/services/impl/CustomerServiceImpl.java
@@ -4,7 +4,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.onlinebanking.core.businesslogic.services.CustomerService;
 import org.onlinebanking.core.dataaccess.dao.interfaces.CustomerDAO;
-import org.onlinebanking.core.domain.exceptions.DAOException;
+import org.onlinebanking.core.domain.exceptions.ServiceException;
 import org.onlinebanking.core.domain.exceptions.EntityNotFoundException;
 import org.onlinebanking.core.domain.exceptions.FailedCustomerRegistrationException;
 import org.onlinebanking.core.domain.models.Customer;
@@ -14,7 +14,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import javax.persistence.NoResultException;
-import javax.persistence.PersistenceException;
 
 @Service
 public class CustomerServiceImpl implements CustomerService {
@@ -31,49 +30,66 @@ public class CustomerServiceImpl implements CustomerService {
     @Transactional
     @Override
     public Customer registerCustomer(Customer customer) {
+        if (customer == null || customer.getTaxPayerId() == null) {
+            throw new ServiceException();
+        }
+
         String taxPayerId = customer.getTaxPayerId();
         try {
             findByTaxPayerId(taxPayerId);
             throw new FailedCustomerRegistrationException(
                     String.format(FAILED_CUSTOMER_REGISTRATION_EXCEPTION_MESSAGE, taxPayerId));
-        } catch (EntityNotFoundException ignored) {}
+        } catch (EntityNotFoundException ignored) {
+        }
 
         try {
             return customerDAO.save(customer);
         } catch (Exception e) {
             logger.error(e);
-            throw new DAOException();
+            throw new ServiceException();
         }
     }
 
     @Transactional
     @Override
     public void deleteCustomer(Customer customer) {
+        if (customer == null) {
+            throw new ServiceException();
+        }
+
         try {
             customerDAO.delete(customer);
         } catch (Exception e) {
             logger.error(e);
-            throw new DAOException();
+            throw new ServiceException();
         }
     }
 
     @Transactional
     @Override
     public Customer updateCustomer(Customer customer) {
+        if (customer == null) {
+            throw new ServiceException();
+        }
         try {
             return customerDAO.update(customer);
         } catch (Exception e) {
             logger.error(e);
-            throw new DAOException();
+            throw new ServiceException();
         }
     }
 
     @Transactional
     @Override
     public boolean assignCustomerToUser(Customer customer, User user) {
+        if (customer == null || user == null) {
+            throw new ServiceException();
+        }
+
         if (customer.getUser() != null) {
             return false;
         }
+
         customer.setUser(user);
         updateCustomer(customer);
         return true;
@@ -82,6 +98,10 @@ public class CustomerServiceImpl implements CustomerService {
     @Transactional(readOnly = true)
     @Override
     public Customer findByTaxPayerId(String taxPayerId) {
+        if (taxPayerId == null) {
+            throw new ServiceException();
+        }
+
         try {
             return customerDAO.findByTaxPayerId(taxPayerId);
         } catch (NoResultException e) {
@@ -89,13 +109,17 @@ public class CustomerServiceImpl implements CustomerService {
                     String.format(ENTITY_NOT_FOUND_EXCEPTION_MESSAGE, " taxPayerId " + taxPayerId));
         } catch (Exception e) {
             logger.error(e);
-            throw new DAOException();
+            throw new ServiceException();
         }
     }
 
     @Transactional
     @Override
     public Customer findByUser(User user) {
+        if (user == null) {
+            throw new ServiceException();
+        }
+
         try {
             return customerDAO.findByUser(user);
         } catch (NoResultException e) {
@@ -103,7 +127,7 @@ public class CustomerServiceImpl implements CustomerService {
                     String.format(ENTITY_NOT_FOUND_EXCEPTION_MESSAGE, " user " + user.getEmail()));
         } catch (Exception e) {
             logger.error(e);
-            throw new DAOException();
+            throw new ServiceException();
         }
     }
 }
